backend/requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
Flask==3.0.2
Flask-Cors==4.0.0
google-ai-generativelanguage==0.4.0
google-api-core==2.24.2
google-auth==2.40.2
google-cloud-speech==2.25.1
google-generativeai==0.3.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.62.3
gunicorn==23.0.0
hume==0.4.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
proto-plus==1.26.1
protobuf==4.25.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.31.0
rsa==4.9.1
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
websockets==10.4
Werkzeug==3.0.1
zipp==3.21.0
